requirements.txt
pandas>=1.0
scikit-learn>=1.0
nltk>=3.5
joblib>=1.0
requests>=2.20
newspaper3k>=0.2.8
beautifulsoup4>=4.9.3
lxml>=4.6
sentence-transformers>=2.2.2
spacy>=3.0
numpy<2
Flask>=2.0
Flask-CORS>=3.0